+++ src/main/java/dev/orne/config/ConfigurableProperty.java
@@ -1,0 +1,27 @@
+/**
+ * 
+ */
+package dev.orne.config;
+
+import static java.lang.annotation.ElementType.FIELD;
+import static java.lang.annotation.RetentionPolicy.RUNTIME;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.Retention;
+import java.lang.annotation.Target;
+
+@Documented
+@Retention(RUNTIME)
+@Target(FIELD)
+/**
+ * Marks a property as automatically configurable.
+ * 
+ * @version 1.0
+ * @author (w) Iker Hernaez<i.hernaez@hif-soft.net>
+ * @since 1.0, 2019-07
+ */
+public @interface ConfigurableProperty {
+
+	/** The configuration key to apply to the property. */
+	String value();
+}